pyproject.toml
# python -m pip install .

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
authors = [{name = "Jarosław Wierzbowski", email = "jaroslawierzbowski2001@gmail.com"}]
dependencies = [
  "chess >= 1.11.1",
  "pytz >= 2025.1",
  "flet >= 0.26.0",
  "opencv-python >= 4.11.0.86",
  "vidgear >= 0.3.3",
  "ruff >= 0.9.6",
  "winaccent >= 2.0.1",
  "pololu-maestro @git+https://github.com/austin-bowen/pololu-maestro.git",
]
dynamic = ["version"]
license = {text = "MIT License"}
name = "kawachess"
requires-python = ">=3.13.2"

[tool.ruff]
exclude = [
  "KAWASAKI",
  ".git",
  "piece_svg",
  "chess.db",
  ".gitignore",
  ".svg",
  ".txt",
  ".pyc",
]
fix = true
indent-width = 4
line-length = 160
target-version = "py313"

[tool.ruff.lint]
ignore = [
  "W191",
  "FURB101",
  "FURB103",
  "FURB177",
  "TRY400",
  "PLE1205",
  "D100",
  "CPY001",
  "D102",
  "D107",
  "D101",
  "D105",
  "D106",
  "D103",
  "D104",
  "D203",
  "D212",
  "COM812",
]
preview = true
select = ["ALL"]

fixable = ["ALL"]

unfixable = ["B", "T201", "T203", "F401"]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
order-by-type = true

[tool.setuptools]
py-modules = []
